Rename Controllers to Controller, make default view
